rename ffi -> utility
